Fix typos and move tests to seperate file
